+++ examples/stoch_distr/stoch_distr.py
@@ -3,8 +3,6 @@ import pyomo.environ as pyo
 import mpisppy.utils.sputils as sputils
 import numpy as np
 import re
-import mpisppy.scenario_tree as scenario_tree
-
 
 # In this file, we create a stochastic (linear) inter-region minimal cost distribution problem.
 # Our data, gives the constraints inside each in region in region_dict_creator
@@ -234,20 +232,8 @@ def region_dict_creator(admm_subproblem_name): #in this precise example region_d
     return region_dict
 
 
-def _scenario_number(stoch_scenario_name, demand=None):
-    if demand == None:
-        scennum = sputils.extract_num(stoch_scenario_name)
-    else: # 3-stage
-        production_num = int(re.search(r'\d+', stoch_scenario_name).group())
-        if demand == "high":
-            scennum = (production_num-1)*2
-        else:
-            scennum = (production_num-1)*2+1
-    return scennum
-
-
 ###Creates the model when local_dict is given, local_dict depends on the subproblem
-def min_cost_distr_problem(local_dict, stoch_scenario_name, cfg, demand=None, sense=pyo.minimize):
+def min_cost_distr_problem(local_dict, stoch_scenario_name, cfg, sense=pyo.minimize):
     """ Create an arcs formulation of network flow for the region and stochastic scenario considered.
 
     Args:
@@ -260,8 +246,7 @@ def min_cost_distr_problem(local_dict, stoch_scenario_name, cfg, demand=None, se
         model (Pyomo ConcreteModel) : the instantiated model
     """
     # Helps to define pseudo randomly the percentage of loss at production
-    scennum = _scenario_number(stoch_scenario_name, demand=demand)
-
+    scennum = sputils.extract_num(stoch_scenario_name)
     # Assert sense == pyo.minimize, "sense should be equal to pyo.minimize"
     # First, make the special In, Out arc lists for each node
     arcsout = {n: list() for n in local_dict["nodes"]}
@@ -269,18 +254,6 @@ def min_cost_distr_problem(local_dict, stoch_scenario_name, cfg, demand=None, se
     for a in local_dict["arcs"]:
         arcsout[a[0]].append(a)
         arcsin[a[1]].append(a)
-
-    # Creating the random demand based on the supply if it is a three-stage example. In which case the original
-    # local_dict["supply"][n] is the average of demand for a node n 
-    if demand is not None:
-        if demand == "high":
-            factor = 1.2
-        elif demand == "low":
-            factor = 0.8
-        else:
-            raise RuntimeError (f"for the 3-stage problem demand should be 'high' or 'low', but it is {demand=}")
-        for n in local_dict["buyer nodes"]:
-            local_dict["supply"][n] *= factor
 
     model = pyo.ConcreteModel(name='MinCostFlowArcs')
     def flowBounds_rule(model, i,j):
@@ -303,26 +276,14 @@ def min_cost_distr_problem(local_dict, stoch_scenario_name, cfg, demand=None, se
     
     model.y = pyo.Var(local_dict["nodes"], bounds=slackBounds_rule)
 
-    if demand is not None: # 3-stage example, defining exportation
-        def export_bounds(m,i):
-            return (0,10)
-        model.export = pyo.Var(local_dict["buyer nodes"], bounds=export_bounds)
-
     model.FirstStageCost = pyo.Expression(expr=\
                     sum(local_dict["production costs"][n]*(local_dict["supply"][n]-model.y[n]) for n in local_dict["factory nodes"]))
     
-    if demand is not None: # 3-stage example
-        model.SecondStageCost = pyo.Expression(expr=\
-                        sum(-810*model.export[n] for n in local_dict["buyer nodes"]))
-    
-    model.LastStageCost = pyo.Expression(expr=\
+    model.SecondStageCost = pyo.Expression(expr=\
                     sum(local_dict["flow costs"][a]*model.flow[a] for a in local_dict["arcs"]) \
                     + sum(local_dict["revenues"][n]*(local_dict["supply"][n]-model.y[n]) for n in local_dict["buyer nodes"]))
 
-    if demand is not None: # 3-stage example
-        model.MinCost = pyo.Objective(expr=model.FirstStageCost + model.SecondStageCost +model.LastStageCost, sense=sense)
-    else: # 2-stage example
-        model.MinCost = pyo.Objective(expr=model.FirstStageCost +model.LastStageCost, sense=sense)
+    model.MinCost = pyo.Objective(expr=model.FirstStageCost + model.SecondStageCost, sense=sense)
     
     def FlowBalance_rule(m, n):
         #we change the definition of the slack for target dummy nodes so that we have the slack from the source and from the target equal
@@ -339,10 +300,6 @@ def min_cost_distr_problem(local_dict, stoch_scenario_name, cfg, demand=None, se
             return sum(m.flow[a] for a in arcsout[n])\
             - sum(m.flow[a] for a in arcsin[n])\
             == (local_dict["supply"][n] - m.y[n]) * min(1,max(0,1-np.random.normal(cfg.spm,cfg.cv)/100)) # We add the loss
-        elif n in local_dict["buyer nodes"] and demand is not None: # for a 3-stage scenario the export changes the flow balance
-            return sum(m.flow[a] for a in arcsout[n])\
-            - sum(m.flow[a] for a in arcsin[n])\
-            + m.y[n] == local_dict["supply"][n] - m.export[n]
         else:
             return sum(m.flow[a] for a in arcsout[n])\
             - sum(m.flow[a] for a in arcsin[n])\
@@ -366,57 +323,12 @@ def scenario_denouement(rank, admm_stoch_subproblem_scenario_name, scenario):
     scenario.flow.pprint()
     print(f"slack values for {admm_stoch_subproblem_scenario_name=} at {rank=}")
     scenario.y.pprint()
-    if hasattr(scenario,"export"): # ie 3-stage problem
-        print(f"export values for {admm_stoch_subproblem_scenario_name=} at {rank=}")
-        scenario.export.pprint()
-
-
-# Only if 3-stage problem for the example
-def MakeNodesforScen(model, BFs, scennum, local_dict):
-    """ Make just those scenario tree nodes needed by a scenario.
-        Return them as a list.
-        NOTE: the nodes depend on the scenario model and are, in some sense,
-              local to it.
-        Args:
-            BFs (list of int): branching factors
-    """
-    ndn = "ROOT_"+str((scennum-0) // BFs[1]) # scennum is 0-based, 
-    # In a more general way we should divide by: BFs[1:]
-    #objfunc = pyo.Expression(expr=0) # Nothing is done so there is no cost
-    retval = [scenario_tree.ScenarioNode("ROOT",
-                                         1.0,
-                                         1,
-                                         model.FirstStageCost,
-                                         [model.y[n] for n in  local_dict["factory nodes"]],
-                                         model),
-              scenario_tree.ScenarioNode(ndn,
-                                         1.0/BFs[0],
-                                         2,
-                                         model.FirstStageCost, 
-                                         [model.export[n] for n in  local_dict["buyer nodes"]], # No consensus_variable for now to simplify the model
-                                         model,
-                                         parent_name="ROOT")
-              ]
-    return retval
-
-
-def branching_factors_creator(num_stoch_scens, num_stage):
-    # BFs is only used with multiple stages
-    if num_stage == 3:
-        # There should always be a high demand and a low demand version
-        assert num_stoch_scens % 2 == 0, f"there should be an even number of stochastic scenarios for this 3 stage problem, but it is odd"
-        BFs = [num_stoch_scens//2, 2]
-    elif num_stage == 2:
-        BFs = None # No need to use branching factors (or even to define them with two stage problems)
-    else:
-        raise RuntimeError (f"the example is only made for 2 or 3-stage problems, but {num_stage} were given")
-    return BFs
 
 
 ###Creates the scenario
 def scenario_creator(admm_stoch_subproblem_scenario_name, **kwargs):
     """Creates the model, which should include the consensus variables. \n
-    However, this function shouldn't attach the consensus variables for the admm subproblems as it is done in admm_ph.
+    However, this function shouldn't attach the consensus variables for the admm subproblems as it is done in admmWrapper.
 
     Args:
         admm_stoch_subproblem_scenario_name (str): the name given to the admm problem for the stochastic scenario. \n
@@ -436,31 +348,17 @@ def scenario_creator(admm_stoch_subproblem_scenario_name, **kwargs):
 
     # Adding dummy nodes and associated features
     local_dict = dummy_nodes_generator(region_dict, inter_region_dict)
+    # Generating the model
+    model = min_cost_distr_problem(local_dict, stoch_scenario_name, cfg)
 
-    # Generating the model according to the number of stages and creatung the tree
-    if cfg.num_stage == 2:
-        model = min_cost_distr_problem(local_dict, stoch_scenario_name, cfg)
-        sputils.attach_root_node(model, model.FirstStageCost, [model.y[n] for n in  local_dict["factory nodes"]])
-    else:
-        BFs = branching_factors_creator(cfg.num_stoch_scens, cfg.num_stage)
-
-        if stoch_scenario_name.endswith('high'):
-            demand = "high"
-        elif stoch_scenario_name.endswith('low'):
-            demand = "low"
-        else:
-            raise RuntimeError (f"the stochastic scenario name should end with 'high' or 'low', but it is {stoch_scenario_name}")
-        model = min_cost_distr_problem(local_dict, stoch_scenario_name, cfg, demand=demand)
-        scennum = _scenario_number(stoch_scenario_name, demand=demand)
-        model._mpisppy_node_list = MakeNodesforScen(model, BFs, scennum, local_dict)
-        model._mpisppy_probability = 1/cfg.num_stoch_scens
+    sputils.attach_root_node(model, model.FirstStageCost, [model.y[n] for n in  local_dict["factory nodes"]])
     
     return model
 
 
-def consensus_vars_creator(admm_subproblem_names, stoch_scenario_name, kwargs, num_stage=2):
+def consensus_vars_creator(admm_subproblem_names, stoch_scenario_name, kwargs):
     """The following function creates the consensus_vars dictionary thanks to the inter-region dictionary. \n
-    This dictionary has redundant information, but is useful for admm_ph.
+    This dictionary has redundant information, but is useful for admmWrapper.
 
     Args:
         admm_subproblem_names (list of str): name of the admm subproblems (regions) \n
@@ -485,12 +383,12 @@ def consensus_vars_creator(admm_subproblem_names, stoch_scenario_name, kwargs, n
         #adds dummy_node in the source region
         if not region_source in consensus_vars: #initiates consensus_vars[region_source]
             consensus_vars[region_source] = list()
-        consensus_vars[region_source].append((vstr,num_stage))
+        consensus_vars[region_source].append((vstr,2))
 
         #adds dummy_node in the target region
         if not region_target in consensus_vars: #initiates consensus_vars[region_target]
             consensus_vars[region_target] = list()
-        consensus_vars[region_target].append((vstr,num_stage))
+        consensus_vars[region_target].append((vstr,2))
     # now add the parents. It doesn't depend on the stochastic scenario so we chose one and
     # then we go through the models (created by scenario creator) for all the admm_stoch_subproblem_scenario 
     # which have this scenario as an ancestor (parent) in the tree
@@ -506,7 +404,7 @@ def consensus_vars_creator(admm_subproblem_names, stoch_scenario_name, kwargs, n
     return consensus_vars
 
 
-def stoch_scenario_names_creator(num_stoch_scens, num_stage=2):
+def stoch_scenario_names_creator(num_stoch_scens):
     """Creates the name of every stochastic scenario.
 
     Args:
@@ -515,10 +413,7 @@ def stoch_scenario_names_creator(num_stoch_scens, num_stage=2):
     Returns:
         list (str): the list of stochastic scenario names
     """
-    if num_stage == 3:
-        return [f"StochasticScenario{i+1}_{demand}" for i in range(num_stoch_scens//2) for demand in ["high","low"]]
-    else:
-        return [f"StochasticScenario{i+1}" for i in range(num_stoch_scens)]
+    return [f"StochasticScenario{i+1}" for i in range(num_stoch_scens)]
 
 
 def admm_subproblem_names_creator(num_admm_subproblems):
@@ -567,10 +462,9 @@ def split_admm_stoch_subproblem_scenario_name(admm_stoch_subproblem_scenario_nam
     """
     # Method specific to our example and because the admm_subproblem_name and stoch_scenario_name don't include "_"
     splitted = admm_stoch_subproblem_scenario_name.split('_')
-    # The next line would require adding the argument num_stage and transferring it to stoch_admm_ph
-    #assert (len(splitted) == num_stage+2), f"appart from the 'ADMM_STOCH_' prefix, underscore should only separate stages"
+    assert (len(splitted) == 4), f"no underscore should be attached to admm_subproblem_name nor stoch_scenario_name"
     admm_subproblem_name = splitted[2]
-    stoch_scenario_name = '_'.join(splitted[3:])
+    stoch_scenario_name = splitted[3]
     return admm_subproblem_name, stoch_scenario_name
 
 
@@ -625,9 +519,3 @@ def inparser_adder(cfg):
                       description="initial seed for generating the loss",
                       domain=int,
                       default=0)
-    
-    cfg.add_to_config("num_stage",
-                      description="choice of the number of stages for the example, by default two-stage",
-                      domain=int,
-                      default=2)
-
