+++ mpisppy/tests/examples/stoch_distr/stoch_distr_admm_cylinders.py
@@ -2,10 +2,14 @@
 # This software is distributed under the 3-clause BSD License.
 # general example driver for stoch_distr with cylinders
 
+# This file is used for the example! It is slightly
+# different from the one in the examples outside of the tests
+# as it allows to consider a three stage example
+
 # Driver file for stochastic admm
 import mpisppy.utils.stoch_admmWrapper as stoch_admmWrapper
 # import stoch_distr ## It is necessary for the test to have the full direction
-import stoch_distr
+import mpisppy.tests.examples.stoch_distr.stoch_distr as stoch_distr
 import mpisppy.cylinders
 
 from mpisppy.spin_the_wheel import WheelSpinner
@@ -15,7 +19,7 @@ import mpisppy.utils.cfg_vanilla as vanilla
 from mpisppy import MPI
 global_rank = MPI.COMM_WORLD.Get_rank()
 
-write_solution = True
+write_solution = False
 
 def _parse_args():
     # create a config object and parse
@@ -48,19 +52,21 @@ def _count_cylinders(cfg):
     return count
 
 
-def _make_admm(cfg, n_cylinders,verbose=None,BFs=None):
+def _make_admm(cfg, n_cylinders,verbose=None):
     options = {}
+    
+    BFs = stoch_distr.branching_factors_creator(cfg.num_stoch_scens, cfg.num_stage)
 
     admm_subproblem_names = stoch_distr.admm_subproblem_names_creator(cfg.num_admm_subproblems)
-    stoch_scenario_names = stoch_distr.stoch_scenario_names_creator(num_stoch_scens=cfg.num_stoch_scens)
+    stoch_scenario_names = stoch_distr.stoch_scenario_names_creator(num_stoch_scens=cfg.num_stoch_scens, num_stage=cfg.num_stage)
     all_admm_stoch_subproblem_scenario_names = stoch_distr.admm_stoch_subproblem_scenario_names_creator(admm_subproblem_names,stoch_scenario_names)
     split_admm_stoch_subproblem_scenario_name = stoch_distr.split_admm_stoch_subproblem_scenario_name
     
     scenario_creator = stoch_distr.scenario_creator
     scenario_creator_kwargs = stoch_distr.kw_creator(cfg)
-    stoch_scenario_name = stoch_scenario_names[0] # choice of any scenario
-    consensus_vars = stoch_distr.consensus_vars_creator(admm_subproblem_names, stoch_scenario_name, scenario_creator_kwargs)
-    admm = stoch_admmWrapper.STOCH_ADMM_PH(options,
+    stoch_scenario_name = stoch_scenario_names[0] # choice of any scenario to create consensus_vars
+    consensus_vars = stoch_distr.consensus_vars_creator(admm_subproblem_names, stoch_scenario_name, scenario_creator_kwargs, num_stage=cfg.num_stage)
+    admm = stoch_admmWrapper.Stoch_AdmmWrapper(options,
                            all_admm_stoch_subproblem_scenario_names,
                            split_admm_stoch_subproblem_scenario_name,
                            admm_subproblem_names,
@@ -71,7 +77,7 @@ def _make_admm(cfg, n_cylinders,verbose=None,BFs=None):
                            mpicomm=MPI.COMM_WORLD,
                            scenario_creator_kwargs=scenario_creator_kwargs,
                            verbose=verbose,
-                           BFs=None,
+                           BFs=BFs,
                            )
     return admm, all_admm_stoch_subproblem_scenario_names
     
@@ -148,8 +154,7 @@ def _wheel_creator(cfg, n_cylinders, scenario_creator, variable_probability, all
 
 
 
-def main():
-    cfg = _parse_args()
+def main(cfg):
     assert cfg.fwph_args is not None, "fwph does not support variable probability"
 
     if cfg.default_rho is None: # and rho_setter is None
@@ -176,4 +181,5 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    cfg = _parse_args()
+    main(cfg)
